requirements.txt
Django==1.10.4
django-bootstrap-form==3.3
django-contrib-comments==1.8.0
django-tinymce==2.6.0
oauthlib==2.0.2
olefile==0.44
Pillow==4.2.1
pycrypto==2.6.1
pytz==2017.2
whitenoise == 3.3.0
gunicorn == 19.7.1
dj-database-url == 0.2.1
python-decouple==3
dj-static==0.0.6
psycopg2==2.6.1